requirements.txt
playwright>=1.43.0
openpyxl>=3.1.0
python-dotenv>=1.0.0